lib/strings: exclude question mark from derivation names
The question mark is a shell metacharacter and may be interpretted by
the shell, and most of nixpkgs expects a derivation output to be safe to
use unquoted.

Realistically the likelihood that this will come up as a real issue is
minimal-to-none, but this strengthens the guarantee that nixpkgs stdenv
derivation store paths are safe unquoted.