fixGPRBuildDarwinDylibNames: init setup hook
The differences between this hook and the existing fixDarwinDylibNames
are:
- it rewrites dylib dependencies of the form `@rpath/libfoo.dylib`
  instead of `libfoo.dylib`, as the former is what results in projects
  built with GPRBuild on Darwin;
- it does so in Mach-O executables as well as dylibs; and
- it tries to be a bit faster and safer regarding bash pitfalls.

Writing a separate hook allows us to avoid an immediate mass rebuild of
Darwin packages, and initially use the hook only on GPRBuild packages
where the existing fixDarwinDylibNames hook does not work.  Ideally,
this hook would be incorporated into fixDarwinDylibNames in future, and
cease to exist separately.