CODEOWNERS: Drop python-modules and python-packages ownership
We currently have 5800 python modules in that path, and we're seeing
roughly 1000 pull requests for these modules per month.

Having @FRidh and @jonringer requested on every PR that touches a
python package isn't helping anyone, nor is it sustainable for any one
person to have that number of incoming notifications.

I think it's time to get rid of that code ownership.