setup-hooks/reproducible-builds.sh: NIX_OUTPATH_USED_AS_RANDOM_SEED
For reproducibility, nixpkgs sets `-frandom-seed` to the first 10
characters of the outpath of the derivation being built.

This PR allows to manually select from which outpath the
10-character prefix is taken, by setting
`NIX_OUTPATH_USED_AS_RANDOM_SEED` as a derivation attribute (or
directly as an environment variable).

For an example use of this functionality, see:

  https://github.com/NixOS/nixpkgs/pull/209870