libopenshot: improve Python dir specification to fix Darwin
Previously, the Python path selection logic was patched in the CMake
files. This patch no longer applies.

Fortunately there is a new way to directly specify the path without
patches. Use that instead on all platforms to fix the CMake logic being
broken on Darwin which results in this module not being detected and
the main application failing to start.