flatcam: mark broken
Tests have been failing since Shapely was updated to 2.0.0, e.g.:

	Traceback (most recent call last):
	  File "/build/source/tests/test_paint.py", line 89, in test_jump
	    result = Geometry.paint_connect(mkstorage(deepcopy(paths)), self.boundary, tooldia)
	  File "/build/source/camlib.py", line 792, in paint_connect
	    candidate.coords = list(candidate.coords)[::-1]
	AttributeError: can't set attribute 'coords'

The AttributeError comes from non-test code, so I assume the program
is actually broken.

Upstream is dead.  There appears to be a maintained fork[1], but it
fails to build with Python 3.9 so I didn't get very far with it[2].
Not dropping the package though, in case it gets fixed.

[1]: https://bitbucket.org/marius_stanciu/flatcam_beta/
[2]: https://bitbucket.org/jpcgt/flatcam/issues/625/makefile-on-beta-branch-thinks-python-311