maintainers: Fix github account names/ids
cust0dian is now known on github as cust0dian-old.
davsanchez is now known on github as DavSanchez.
luxferresum is now known on github as luxzeitlos.
benesim is now known on github as BeneSim.
DPDmancul has no github handle
maxwell-lt is now known on github as Maxwell-lt.
paddygord is now known on github as avaunit02.
ewok is now known on github as ewok-old.
nrhelmi is now known on github as NRHelmi.
elnudev is now known on github as ElnuDev.
jkarlson is now known on github as ethorsoe.
Zebreus is now known on github as zebreus.
vtrf is now known on github as ratsclub.
loveisgrief has no github handle
sno2wman is now known on github as SnO2WMaN.
kidsan is now known on github as Kidsan.
harrisonthorne is now known on github as muni-corn.
teozkr is now known on github as nightkr.
0xb10c is now known on github as 0xB10C.
NekomimiScience is now known on github as ScarletHg.
tonyshkurenko is now known on github as antonshkurenko.
notbandali is now known on github as bandali0.
Shrimpram is now known on github as shrimpram.
munksgaard is now known on github as Munksgaard.
MonAaraj is now known on github as ribosomerocker.
uthar is now known on github as Uthar.