python310Packages.raincloudy: Fix build
Due to an improper update during a python-updates cycle this package
broke.

With version 1.2.0 there is now `raincloudy.aio`, which does not get
installed on recent setuptools versions, due to a misconfigured packages
configuration in setup.py.

Also convert the build process to a PEP517 based build using
setuptools and fix the distinfo version by adding setuptools-scm.

Additionally `aiohttp` is an undeclared dependencym, for which a PR has
already been submitted upstream.