nix-prefetch-docker: fix cross-platform use
nix-prefetch-docker failed to run when the target Docker image didn't
exist for the host platform, regardless of the target platform specified
for the image.

> $ nix-prefetch-docker --json pihole/pihole 2023.02.01
> FATA[0001] Error parsing manifest for image: choosing image instance: no image found in manifest list for architecture amd64, variant "", OS darwin

The failure occurred because some calls to the skopeo CLI didn't specify
the target OS and architecture. This change fixes the problem by
wrapping every call to skopeo.