linuxkit: Fix Darwin runtime OS detection
The wrapper that linuxkit uses around the macOS Virtualization framework
tries to detect which OS it is running on, and aborts if it thinks that
the macOS version is not new enough to have the required framework
features.  This commit (mostly) fixes that macOS version detection when
the code is linked as if it is linked against the 10.12 SDK (as the
current Darwin ld wrapper does via the -platform_version switch), but is
then run on macOS 11 or later.