nixos/nginx: add prependConfig option
Nginx allows us to load dynamic modules, however the `load_module` directive
must be close to the top of the config so that dependent directives can be
processed correctly.

This change adds a `prependConfig` option which reflects the function of the
existing `appendConfig` option. A new option is neccessary because the `config`
option is mutually exclusive with some other options which users may not want to
reimplement.

Note that it is currently possible to get around this using something like:

```
logError = ''stderr
load_module /path/to/shared/object.so'';
```

However this is clunky and may be prone to breaking in the future.