nixos/tests/containers-next/wireguard: init
This... kinda escalated. Context: I disliked the current networking
setup I have using RFC108 containers since it's entirely static, i.e.
without making use of DHCP/RA from the new module.

This was an attempt to draft an alternative solution. I.e. to create a
VPN tunnel between a bunch of servers and give a subnet of that to a
virtual zone containing several containers. Address delegation from the
subnet's pool is taken care of by the module automatically, so one can
easily spin up new containers without much effort (which is especially
useful with improved support for imperative containers).

The only problem was that I wanted to have DNS for each container.
Normally that's done via `nss-mymachines(8)`, however it only applies to
the container's host. Also, resolved is orthogonal to that, the NSS
lookups defined in `/etc/nsswitch.conf` usually invoke `nss-resolve(8)`
for DNS from resolved and then mymachines for machined.

Since the addresses are exposed via DBus, they can be obtained by a DBus
query which is now connected using some glue code written in Lua to
kresd which serves as DNS inside the VPN.