buildDotnetModule: Don't use the --output argument
In SDK version 7.0.200 only, the `dotnet pack` and `dotnet publish`
commands will fail if used with a solution file (as opposed to a single
project file). This breaking change will be reverted in the next update.

Until then, specify the "PublishDir" or "PackageOutputPath" properties
directly as a workaround.

Reference: https://github.com/dotnet/sdk/issues/30624#issuecomment-1432118204