nixos/systemd-shutdown: use prebuilt shutdown ramfs derivation
I changed the shutdown ramfs to be built alongside the system config
rather than be built on-the-fly at shutdown. See discussion starting at
https://github.com/NixOS/nixpkgs/pull/185085#issuecomment-1368503199 for
details on motivation, but basically there's not much reason to only do
this at shutdown (even if it's quick to run), dracut does not do this,
and it requires building a cross rustc toolchain for make-initrd-ng even
if no other Rust programs exist on the target image just to run the one
script.

This change technically could be breaking, but only if the user manually
made a unit that dependend on `generate-shutdown-ramfs.service`, which
doesn't seem like a use-case that would give someone any benefit anyway
with the options available in the module. I can change the service name
back to `generate` if someone feels otherwise that this will be a real
problem.

I opted not to unpack the booted initrd image for shutdown because
unpacking it is a hard problem, and that would ignore the options
already available in this module (and are already used at least by the
ZFS module and some people's configs).

It would be nice if we could try to use the same systemd version and
other software versions that we booted with (just to ensure no issues),
but that would involve modifying system top-level build and loading from
something like `/run/booted-system/shutdownramfs`, which would on the
whole be more complicated than the potential benefit I feel it would
give.