android-ndk: allow -no-pie
We add -pie if we don't find -shared in the arguments. However this does not
work in the presense of @response files.  We also do not respect -no-pie, and
add -pie regardless. This improves the argument parsing, and allows to prevent
the forced -pie flag.

(cherry picked from commit bd4cdf346c0e43f56b1055fc649b9e9987acd601)