etesync-dav: fix broken dependancies
Related #206852

This was failing due to newer versions of Flask requiring newer versions
of werkzeug. I referenced:

https://github.com/etesync/etesync-dav/blob/master/requirements.txt

to downgrade flake and use the current wtforms. The issue with using
flask-wtf==0.15.1 and wtforms==3.0.1 is there are tests in flask-wtf
that reference import paths that are removed in wtforms 3.0.0:

https://github.com/wtforms/flask-wtf/blob/v0.15.1/tests/test_form.py#L5

So I've just skipped the tests that failed. Once everything compiled I
tried running the binary which complained:

ModuleNotFoundError: No module named 'pkg_resources'

So I added setuptools to the propagatedBuildInputs. After I was able to
render the page at http://localhost:37358/.web/add/