nixos/containers-next: remove `rendered` option
This was a hack to be able to introspect the config of a container e.g.
to find prometheus exporters, similar to how `resources.machines` in
nixops used to work. However, this is kinda dirty and can be solved in a
cleaner manner, i.e. by mapping the `system-config` option of each
declarative container to the evaluated NixOS system.