matrix-conduit: use importCargoLock instead of hacky patch
As `cargo vendor` currently doesn't support workspace inherited manifest
values, a patch that replaced the values manually was used as a workaround.

Now that importCargoLock has support for this, and we're probably going
to move towards it anyways, let's switch to it now, as there's a clear
benefit either way.

This also switches to using our copy of SQLite instead of the vendored
one included in `libsqlite3-sys` -- because it's preferred, but also
because it causes the build to fail, now that dependencies are read-only.