prowlarr: 1.2.0.2583 -> 1.2.2.2699
https://github.com/Prowlarr/Prowlarr/releases/tag/v1.2.2.2699

This update adds support for aarch64-darwin systems.

Refactors the url download logic by relying strictly on the system
identifier and using a unified throw that is more informative.

Also prowlarr now seems to be doing releases on the master branch,
instead of develop. I updated the update script accordingly.