nixos/tests/home-assistant: Overhaul and refactor
Prepare the tests for a change in dependency handling, by not relying on
bespoke files dropped into the package output.

Instead we now check the journal log for whether a configured component
was setup, once for the initial specialisation another time for the one
introducing esphome configuration.

Also improve abstractions for getting journal data relative to a cursor
and generally make a few things more concise.