mkDerivation: Keep legacy name semantics
This avoids a mass-rebuild as many expressions currently use the
name+version pair instead of just a name.

The old meaning of name is based on the abused "name part" of store
paths, but there is no real reason for the attrset to use this odd
definition of name, except perhaps strong backward compatibility.