gcc: fix libgccjit cross: enablePlugin=true
Before this commit, `pkgsCross.*.libgccjit` would fail to build.
Apparently `--enable-lang=jit` requires `--enable-plugin`, but the
configury is unaware of this requirement.  The failure mode is
difficult to diagnose; the build fails with a compile error due to
`<dlfcn.h>` not being included.

This commit adds `enablePlugin=true` and an `assert` to check for
the unsupported combination and provide feedback at eval time.