gcc: rename enablePlugin to enablePlugins
This commit implements @SuperSandro2000's suggestion that
`enablePlugins` is a better name for this option:

  https://github.com/NixOS/nixpkgs/pull/216237#discussion_r1113597885

This makes a lot of sense.  The option does not enable one specific
plugin, and in fact does not necessarily enable *all* plugins
either.  Rather, it enables the general infrastructure for gcc
plugins.

For example, you can combine `--enable-plugins` with
`--disable-libcc1` (which is the GDB plugin).