nixos/documentation: Link Devhelp files
While Devhelp now supports finding API documentation in `/share/doc` and the new `gi-docgen` toolchain places it there, older projects using gtk-doc (e.g. GLib) still install it to `/share/gtk-doc` and/or `/share/devhelp/books`.

For people using the major DEs this is not problem since they have `/share` in `pathsToLink` but we want to be explicit and make it work for everyone.