gnome.gnome-settings-daemon: 43.0 → 44.beta
The xsettings module will now always set `Gtk/IMModule` key in XSettings to `ibus`
when GTK’s IM module is set to an empty value (the default).
This will produce an incorrect value in GtkSettings for X11 apps but that does not really
matter as GTK will fall back to `gtk-im-context-simple` as before,
when the requested `ibus` module is not present.

https://gitlab.gnome.org/GNOME/gnome-settings-daemon/-/compare/43.0...44.beta

Changelog-Reviewed-By: Jan Tojnar <jtojnar@gmail.com>