anki-bin: set correct name and version for derivation
anki-bin is built with buildFHSUserEnv on Linux, which doesn't set a
version for the resulting derivation.  This commit overrides the
resulting derivation to have a version set correctly.

This is important for end-users to be able to easily tell what
version of Anki they will get when they install `anki-bin`.  It is
normally very important to use the correct version of Anki.

(cherry picked from commit 0467f6d76f4229ce8bba099b3755087c5f6a1c83)