cudaPackages: point nvcc at a compatible -ccbin
This is a hot-fix to un-break cuda-enabled packages (like tensorflow,
jaxlib, faiss, opencv, ...) after the gcc11->gcc12 bump. We should
probably build the whole downstream packages with a compatible stdenv
(such as gcc11Stdenv for cudaPackages_11), but just pointing nvcc at the
right compiler seems to do the trick

We already used this hack for non-redist cudatoolkit. Now we use it more
consistently.

This commit also re-links cuda packages against libstdc++ from the same
"compatible" gcc, rather than the current stdenv. We didn't test if this
is necessary -> need to revise in further PRs.

NOTE: long-term we should make it possible to override -ccbin and use
e.g. clang