Revert "python3.pkgs.cryptography: use openssl_1_1"
This reverts commit 8b1f16573cc6fa50facc3fe3154ba437e64a9220.

Cryptography was pinned to openssl_1_1 to prevent further breakage among
its consumers. Most notably that was pyopenssl, which has improved quite
a bit since then and should not cause problems any longer.

It is therefore time to give this another shot, given that we cannot in
good conscience make such widespread use of openssl 1.1 a reality in the
upcoming NixOS 23.05 release.