onlyoffice-bin: add gcc lib to runtimeInputs
OnlyOffice is currently not launching
(https://github.com/NixOS/nixpkgs/issues/218238), due to a missing
library (libstdc++.so.6).

I used nix-locate to see which outputs provide it, and
`gcc-unwrapped.lib` is the one that makes more sense to me, but I'm not
sure if there's a better way to supply that lib.