nixos/display-managers: add option to disable starting graphical-session
Add a new option to allow to disable starting the
graphical-session.target automatically. This might be useful to disable to
let the window manager starting the session itself rather than letting the
display manager doing it before the window manager is actually started.
Setting this to false can fix issues with systemd services starting too early
on Wayland.

Co-Authored-By: Nicolas Berbiche <nicolas@normie.dev>
Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@proton.ch>