lib/types: allow changing the name of `name`
Adds a new `nameArgument` argument to `submoduleWith` that allows
changing which argument the "name" (last item of `loc`, i.e. attribute
name or option name this submodule appears under) is bound to.

This avoids conflicts when `moduleType` is used inside a submodule,
where `name` would be set to both the outer submodule's name and the
inner one's. If the outer submodule type sets `nameArgument` to
something different, `moduleType` will continue using `name`, avoiding
the conflict.

Ideally, there should probably be a way to change the inner `name` too
to remove all ambiguity.