home-assistant-component-tests: Reduce to two parallel test runners
The component tests are rather short, and each runner has a certain
setup cost. So we're better off, just starting fewer of them and have
them do more work.

It also works around a race condition, that upstream likely isn't
running into, because of the hardware specs¹ for the runners.

[1] https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners#supported-runners-and-hardware-resources