nixos/initrd-network: run udhcpc in the background
initrd networking is often used to provide encryption keys to
non-interactive computers. These computers are likely to be in a
data-center style environment where direct access is difficult. If
dhcp fails for any reason, this can mean a call out to recover the
computer.

To tolerate transient failures, keep the dhcp client running in the
background.