element-{web,desktop}: hack to make ofborg maintainer pings work again
Until now we had the problem that the matrix team wasn't pinged on
element changes because the version data is in a JSON and the position
of the `version`-attribute wrongly pointed to
`element-{web,desktop}.nix`:

    nix-instantiate -E 'with import ./. {}; builtins.unsafeGetAttrPos "version" element-web' --eval
    { column = 22; file = "/home/ma27/Projects/nixpkgs/pkgs/applications/networking/instant-messengers/element/element-web.nix"; line = 24; }

This is a problem because ofborg checks if modified file is part of a
derivation that got changed in a PR. I.e. only pings for element's
maintainers would be added to an element update PR if `pin.json` (which
gets modified in that case) would be recognized as file being a part of
the changed derivations (element-web/element-desktop)[1]

However, JSON imports don't propagate attribute positions (I don't know
how one would that sanely implement btw), so I decided to change
`pin.json` to a `pin.nix` and merge the relevant contents into
element-web/element-desktop.

This is kinda hacky, but as a maintainer I'd like to get modified if
somebody touches element so I can review & merge that.

With this change the position detection works fine now:

    { column = 3; file = "/home/ma27/Projects/nixpkgs/pkgs/applications/networking/instant-messengers/element/pin.nix"; line = 2; }

[1] https://github.com/NixOS/ofborg/blob/released/ofborg/src/maintainers.nix